fixed relationship between Lodge and lodgeType.
fixed relationship between Lodge and lodgeType. Added a factory to add some default types. Also worked a little bit on the searchResultComponent and fixed the communication between it and the datePickerComponent.

+++ project_bali_app/database/migrations/2020_12_13_160716_create_lodge_reviews_table.php
@@ -16,10 +16,10 @@ class CreateLodgeReviewsTable extends Migration
         Schema::create('lodge_reviews', function (Blueprint $table) {
             $table->id();
 
-            $table->integer('user_id');
+            $table->unsignedInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
 
-            $table->integer('lodge_id');
+            $table->unsignedBigInteger('lodge_id');
             $table->foreign('lodge_id')->references('id')->on('lodges')->onDelete('cascade');
 
             $table->text('comment');
