+++ project_bali_app/database/migrations/2020_12_13_162259_create_invoices_table.php
@@ -16,7 +16,7 @@ class CreateInvoicesTable extends Migration
         Schema::create('invoices', function (Blueprint $table) {
             $table->id();
 
-            $table->integer('booking_id');
+            $table->unsignedBigInteger('booking_id');
             $table->foreign('booking_id')->references('id')->on('bookings')->onDelete('cascade');
 
             $table->timestamps();
